+++ ACL/src/ACL.Business/Domain/Entities/Branch.cs
@@ -2,8 +2,8 @@
 
 public partial class Branch
 {
-    public uint Id { get; set; }
-    public uint CompanyId { get; set; }
+    public ulong Id { get; set; }
+    public ulong CompanyId { get; set; }
 
     public string Name { get; set; } = null!;
 
@@ -13,11 +13,11 @@ public partial class Branch
 
     public byte Status { get; set; }
 
-    public uint Sequence { get; set; }
+    public ulong Sequence { get; set; }
 
-    public uint CreatedById { get; set; }
+    public ulong CreatedById { get; set; }
 
-    public uint UpdatedById { get; set; }
+    public ulong UpdatedById { get; set; }
 
     public DateTime CreatedAt { get; set; }
