+++ ACL/src/ACL.Business/Domain/Entities/Country.cs
@@ -2,7 +2,7 @@
 
 public partial class Country
 {
-    public uint Id { get; set; }
+    public ulong Id { get; set; }
 
     public string Name { get; set; } = null!;
 
@@ -12,11 +12,11 @@ public partial class Country
 
     public byte Status { get; set; }
 
-    public uint Sequence { get; set; }
+    public ulong Sequence { get; set; }
 
-    public uint CreatedById { get; set; }
+    public ulong CreatedById { get; set; }
 
-    public uint UpdatedById { get; set; }
+    public ulong UpdatedById { get; set; }
 
     public DateTime CreatedAt { get; set; }
