+++ ACL/src/ACL.Business/Domain/Entities/Module.cs
@@ -2,7 +2,7 @@
 
 public partial class Module
 {
-    public uint Id { get; set; }
+    public ulong Id { get; set; }
 
     public string Name { get; set; } = null!;
